backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-mail==0.9.1
python-dotenv==1.0.0 